+++ CHANGELOG.md
@@ -2,4 +2,9 @@
 
 ## Version 0.1 (development)
 
+> Paths used in this project are meant only for devcontainers. If you are using any other platform, kindly update the paths.
+
 - Updated enviroment.yml file
+- Created configs for `densenet`, `resnet50` & `vgg16`
+- initiallized `dvc` in project
+- added preprocess_data.py and train_model.py scripts